requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
Pillow==2.2.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
Werkzeug==1.0.1
xlrd==1.2.0
